+++ Content.Server/Arcade/BlockGame/BlockGameArcadeSystem.cs
@@ -1,5 +1,5 @@
 using Content.Server.Power.Components;
-using Content.Server.UserInterface;
+using Content.Shared.UserInterface;
 using Content.Shared.Arcade;
 using Robust.Server.GameObjects;
 using Robust.Shared.Player;
