+++ Content.Server/Lock/EntitySystems/ActivatableUIRequiresLockSystem.cs
@@ -1,7 +1,8 @@
 using Content.Server.Lock.Components;
 using Content.Server.Popups;
-using Content.Server.UserInterface;
+using Content.Shared.UserInterface;
 using Content.Shared.Lock;
+using Content.Server.UserInterface;
 
 namespace Content.Server.Lock.EntitySystems;
 public sealed class ActivatableUIRequiresLockSystem : EntitySystem
