+++ Content.Server/Power/EntitySystems/ActivatableUIRequiresPowerSystem.cs
@@ -1,8 +1,9 @@
 using Content.Shared.Popups;
 using Content.Server.Power.Components;
-using Content.Server.UserInterface;
+using Content.Shared.UserInterface;
 using JetBrains.Annotations;
 using Content.Shared.Wires;
+using Content.Server.UserInterface;
 
 namespace Content.Server.Power.EntitySystems;
