+++ Content.Server/Research/TechnologyDisk/Systems/DiskConsoleSystem.cs
@@ -1,6 +1,6 @@
-﻿using Content.Server.Research.Systems;
+using Content.Server.Research.Systems;
 using Content.Server.Research.TechnologyDisk.Components;
-using Content.Server.UserInterface;
+using Content.Shared.UserInterface;
 using Content.Shared.Research;
 using Content.Shared.Research.Components;
 using Robust.Server.Audio;
