+++ Content.Server/Silicons/Borgs/BorgSystem.Ui.cs
@@ -1,5 +1,5 @@
-﻿using System.Linq;
-using Content.Server.UserInterface;
+using System.Linq;
+using Content.Shared.UserInterface;
 using Content.Shared.Database;
 using Content.Shared.NameIdentifier;
 using Content.Shared.PowerCell.Components;
