+++ Content.Server/Sound/EmitSoundSystem.cs
@@ -1,6 +1,6 @@
-﻿using Content.Server.Explosion.EntitySystems;
+using Content.Server.Explosion.EntitySystems;
 using Content.Server.Sound.Components;
-using Content.Server.UserInterface;
+using Content.Shared.UserInterface;
 using Content.Shared.Sound;
 using Robust.Shared.Random;
