+++ Content.Server/Store/Systems/StoreSystem.cs
@@ -1,5 +1,5 @@
 using Content.Server.Store.Components;
-using Content.Server.UserInterface;
+using Content.Shared.UserInterface;
 using Content.Shared.FixedPoint;
 using Content.Shared.Implants.Components;
 using Content.Shared.Interaction;
