+++ Content.Server/UserInterface/ActivatableUISystem.Power.cs
@@ -1,5 +1,6 @@
 using Content.Server.PowerCell;
 using Content.Shared.PowerCell;
+using Content.Shared.UserInterface;
 using Robust.Shared.Containers;
 
 namespace Content.Server.UserInterface;
