+++ Content.Shared/Access/Components/ActivatableUIRequiresAccessComponent.cs
@@ -1,0 +1,11 @@
+using Content.Shared.Access.Systems;
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Access.Components;
+
+[RegisterComponent, NetworkedComponent, Access(typeof(ActivatableUIRequiresAccessSystem))]
+public sealed partial class ActivatableUIRequiresAccessComponent : Component
+{
+    [DataField]
+    public LocId? PopupMessage = "lock-comp-has-user-access-fail";
+}
